+++ nodemon.json
@@ -2,5 +2,5 @@
 	"watch": ["src"],
 	"ext": ".ts,.js",
 	"ignore": [],
-	"exec": "tsx ./src/clusterMod.ts"
+	"exec": "tsx ./src/index.ts"
 }
