+++ src/index.ts
@@ -1,0 +1,4 @@
+import startServer from './server.js';
+
+const port = process.env.PORT || '4000';
+startServer(port);
